bash linking script improved

+++ nixos/scripts/modules/bashlinking.nix
@@ -3,7 +3,14 @@
     system.activationScripts.binbash = {
         deps = [ "binsh" ];
         text = ''
-            ln -s /bin/sh /bin/bash
+            bash_path="/bin/bash"
+
+            if [ -e $bash_path ]
+            then
+                echo "$bash_path already exists"
+            else
+                ln -s /bin/sh /bin/bash
+            fi
         '';
     };
 }
